+++ pkg/metric_exporter/metric_exporter.go
@@ -1,0 +1,9 @@
+package metric_exporter
+
+import (
+	"time"
+)
+
+type MetricExporter interface {
+	Export(dateTime time.Time, projectID, metric, instanceName string, metricPoints []string)
+}
